[TASK] Simplify table handling in `DefaultTCASchema`
With #103015 the table definition merge has been moved
to an earlier point, providing an associative array
with the table definitions.

This change modifies `DefaultTcaSchema` to operate on
that array and simplifies the required handling code.

More precise doc-block annotations for humans and
PHPStan are added along the way.

Resolves: #103016
Related: #103015
Related: #103014
Releases: main
Change-Id: I7d2120ee836854b2fa9072e44444e7387873bd2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82760
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>