[TASK] Use `podman` before `docker` in `Build/Scripts/runTests.sh`
The `Build/Scripts/runTests.sh` supports `podman` and `docker`
as container cli binary since quite a while now, albeit docker
has been kept as default if not specified using `-b`.

This change removes the hard default and will now use `podman`
as default. In case `podman` is not available `docker` is used.
As a side-effect, the binary for CI is no longer enforced the
hard way.

PostgreSQL 16 is added as database version.

Resolves: #103018
Releases: main, 12.4, 11.5
Change-Id: Iedcd7cbcd2380adb6781bd8b4f075da86e25c03b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82761
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>