[TASK] Update locales translation files
TYPO3 uses a composer package [1] to import
locales along with translations and a custom
script has been added to create and update
included translation files.

This change updates the language files with
the last updates and requires the package
with the current highest version as minimum.

Used command(s):

> composer req --dev \
    "sokil/php-isocodes-db-i18n":"^4.0.19"

> php Build/Scripts/updateIsoDatabase.php

[1] sokil/php-isocodes-db-i18n

Resolves: #103027
Related: #100659
Releases: main, 12.4
Change-Id: I2839113ed3894cec58b41c85abd59039cab00725
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82774
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>