[TASK] Add `composer` dispatcher to `runTests.sh`
This change adds a composer command dispatcher to
`Build/Scripts/runTests.sh` to allow the execution
of any composer command without the hassle to have
a proper composer installed on the host system.

This also helps for onboarding new contributors not
having a proper php and composer ecosystem installed
on the host system.

For example, to raise dependencies following can now
be used:

> Build/Scripts/runTests.sh -s composer -- require --dev \
  typo3/testing-framework:dev-main

Other examples:

> Build/Scripts/runTests.sh -s composer -- dumpautoload
> Build/Scripts/runTests.sh -s composer -- info | grep "symfony"

Resolves: #103045
Releases: main, 12.4, 11.5
Change-Id: I65fa14e2c8d0d1a1204b3bed701d8c3458b6e3bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82839
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>