[BUGFIX] Also fetch outdated extensions in extensionmanager
The query for fetching extension versions is extended to
also find "outdated" extension versions (review_state = -2).

Note: For now, this does not change anything in the EM,
since the remote XML does not use the -2 state currently,
due to the wrong constraint. This can however now be
properly fixed in XML.

Resolves: #93082
Releases: main, 12.4, 11.5
Change-Id: I522a2c166c6d1af6e8b2a14de51d63a15c4211d2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82848
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>