[BUGFIX] Wrong language labels in StandardContentPreviewRenderer
The language labels for `noPluginSelected` and `edit` in
StandardContentPreviewRenderer can not be resolved, because the
extension path is wrong.

This change fixes the extension path and additionally streamlines
one label output, by using `htmlspecialchars` to escape the
generated label similar to the other label output in this class.

Additionally a similar wrong extension path in
`ExampleLinkType.rst.txt` has been adapted.

Resolves: #103095
Related: #100675
Releases: main, 12.4
Change-Id: I4882133025c938fc3b6928b36237e782e9465dcc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82884
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>