[TASK] Update container image versions
New core-testing-* images have been build and published in
the TYPO3 testing-infrastructure repository and pulled to
the core ci gitlab runner instances. [1][2]

This change modifies `Build/Scripts/runTests.sh` to set
new image version to ensure the usage for all changes
based on this change.

For information what changed in the image see the patches
in the testing-infrastrucure repository linked below.

* core-testing-php82 => 1.10
* core-testing-php83 => 1.11

[1] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/e6719d303eacf2f377e308c800b474b735145d61
[2] https://github.com/advisories/GHSA-7c6p-848j-wh5h

Resolves: #103096
Releases: main, 13.0, 12.4, 11.5
Change-Id: If9c5ae00334acc4fa5a1417dcbf9f77d31dff62d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82844
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>