[DOCS] Document how to replace a linktype
Since the linktypes are automatically registered in Linkvalidator,
it is no longer quite as easy to replace the "external" linktype.
You cannot register another linktype using the same identifier
"external", so a different identifier must be used.

A minimal example is added to the developer section in the
documentation.

Resolves: #101673
Related: #96935
Releases: main, 12.4
Change-Id: I9afa1a77fe56675dff2dbc1bacc3635b1b9b36c6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82849
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>