[TASK] Update `phpstan/phpstan` version
A new PHPStan version has been releases,
which is pulled in automatically in the
composerInstallMax nightly test.

The higher version allows to remove a
ignore pattern from the baseline and
a superfluous key unset is detected.

This change raises the PHPStan version,
removes the superfluous `unset()` and
regenerates the baseline to remove the
invalid ignore pattern.

Used command(s):

> Build/Scripts/runTests.sh -s composer \
    -- require --dev "phpstan/phpstan":"^1.10.58"
> Build/Scripts/runTests.sh \
    -s phpstanGenerateBaseline

Resolves: #103105
Releases: main, 13.0, 12.4
Change-Id: I689079eaac7f9461084e6a71ff08ef716cf551e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82933
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>