[TASK] Simplify DH->checkRecordUpdateAccess()
Argument 3 $data of `checkRecordUpdateAccess()` is
funny: It is only used if argument 4 is given as
well. This is only done in method `process_datamap()`,
so `checkRecordUpdateAccess` hook method is
only triggered from within that method.

This situation allows us to move the hook call into
`process_datamap()` directly, making it more obvious
it's only triggered by that `main` method. This
allows us to get rid of arguments 3 and 4. Also
point out a logical issue within the method.

Resolves: #103126
Releases: main
Change-Id: I4f4b2711784c005ed312ca491775c3e1d39b35d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83010
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>