[TASK] Update container image versions
New core-testing-* images have been build and published in
the TYPO3 testing-infrastructure repository and pulled to
the core ci gitlab runner instances. [1]

This change modifies `Build/Scripts/runTests.sh` to set
new image version to ensure the usage for all changes
based on this change.

For information what changed in the image see the patches
in the testing-infrastrucure repository linked below.

* core-testing-php82 => 1.11
* core-testing-php83 => 1.12

[1] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/072687df3a90e808143b3eded59176ee593e8112

Resolves: #103143
Releases: main, 13.0, 12.4, 11.5
Change-Id: Ia0ec77beb11e3f24614a0f5b531799494c1af988
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83028
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>