[BUGFIX] Indicate loading process when requesting context menu
This commit changes the rendering behavior of context menus. Previously,
the context menu became visible once the server requests were handled
successfully. This is an issue in scenarios where either the server is
under load and may need some time to send a response, or the network
connectivity of the backend user is not optimal.

Now, when requesting a context menu, it's container is shown before any
data is received and a spinner is rendered.

Resolves: #103197
Releases: main, 12.4
Change-Id: I0cd36976a407bce7ce45520d86d7d28a5719cd92
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83127
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>