[BUGFIX] Streamline `typo3-spinner-icon` element with SVG asset
The Lit element `typo3-backend-spinner` uses a hard-coded SVG string,
which has to be the same as the `spinner-circle` asset file.

However, the SVG string was outdated as it neither respected the
`currentColor`, nor respected the configured icon size. Both issues are
fixed in this commit. Since `currentColor` is now taken into account,
the variant styling has been removed.

Resolves: #103198
Releases: main, 12.4
Change-Id: I24eababb8716076eb4db55a6bcd3eb43e98ea1fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83128
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>