[TASK] Avoid phpunit static TestCase->never() and once()
The methods are no longer static in phpunit 11 and
should not be used in data providers. We split a data
provider to deal with this.

The final switch to not call these statically will
be done when phpunit is raised, since it otherwise
collides with cgl checks and current phpunit 10.

Resolves: #103216
Releases: main
Change-Id: I90c787d1c8b68e0bf02be0e4f056ef6ebb8ca9ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83148
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>