[BUGFIX] Make page tree workspace color important
The page tree can be colored using user TS, example
for page uid '65' and all pages below:

`options.pageTree.backgroundColor.65 = #41afd8`

If a record is changed in workspaces, that page is
also colored. In v11, the 'ws changed' color has a
higher priority than the custom background color,
which is good UX wise. This stopped working in v12
with #97344.

v13 is not affected since the "color approach" has been
changed. The v12 patch is to re-add the "ws changed" color
as "fill" with an !important, to force overriding the
svg inline style of the custom background color, similar
to v11.

Resolves: #103239
Related: #97344
Releases: 12.4
Change-Id: I8e9eee7d85fb2ca18c28247f9fdcb68dc7a9139f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83172
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>