[TASK] Replace GeneralUtility::hmac usage in ext:fluid
With #102761, the new `HashService` has been introduced to ensure
that HMAC generation will always use an additional secret. As a
follow-up patch, `GeneralUtility::hmac` will be deprecated when all
usages in the core have been replaced with the new `HashService`.

With this change, usages of `GeneralUtility::hmac` in ext:fluid are
replaced by the new HashService.

Resolves: #103246
Releases: main
Change-Id: I7090d13590dc0c5d871eadf73aa9c1278cda1f88
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83178
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>