[TASK] Replace GeneralUtility::hmac usage in ext:frontend
With #102761, the new `HashService` has been introduced to ensure
that HMAC generation will always use an additional secret. As a
follow-up patch, `GeneralUtility::hmac` will be deprecated when all
usages in the core have been replaced with the new `HashService`.

With this change, usages of `GeneralUtility::hmac` in ext:frontend
are replaced by the new HashService.

Resolves: #103250
Related: #102761
Related: #103245
Releases: main
Change-Id: I077cf02a170178c750662c8491bf725ce234c036
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83260
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>