[TASK] Replace GeneralUtility::hmac usage in ext:core
With #102761, the new `HashService` has been introduced to ensure
that HMAC generation will always use an additional secret. As a
follow-up patch, `GeneralUtility::hmac` will be deprecated when all
usages in the core have been replaced with the new `HashService`.

With this change, usages of `GeneralUtility::hmac` in ext:core
are replaced by the new HashService.

Resolves: #103248
Related: #102761
Related: #103245
Releases: main
Change-Id: Iccec4e857790916aa53fb2505c340cde4da8cf9b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83295
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>