[TASK] Separation of concerns while rendering Page module
All PageTsConfig options are now in DrawingConfiguration
and named properly.

All plain labels are now moved to Fluid directly via <f:translate>

Random id="{uniqueId}" are removed from markup, reducing
the usages to the AbstractGridObject, which hopefully
will vanish in the near future.

RecordRememberer is a singleton and injected as much as possible.

The "languageMode" is now resolved into a PageViewMode enum.

Resolves: #103345
Releases: main
Change-Id: I78f33fed409db2a1c5528e734ed19ad67aeb4e89
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83322
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>