[TASK] Streamline Backend Layout View code
This change adapts some places around Backend Layouts,
which is a pre-patch in order to centralize
previously used code from 2013 to move towards a more
generic concept, which needs to go into EXT:core
as the structure should also be useful and evaluated
in EXT:frontend.

This change now cleans up places which are non-breaking
but hardens PHP code without changing the underlying
logic.

Resolves: #103365
Releases: main
Change-Id: I77382d93342e5c2e45966f96bf485619c79f25f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83362
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>