[DOCS] Rename .rst includes to .rst.txt
All files ending on .rst are interpreted as documents in both the new and old rendering. This leads to warnings and errors. So all files to be included are renamend to .rst.txt as is standard.

Additionally include files should not include the Includes.rst.txt as it is already included in the main file. Those includes are removed.

Resolves: #103353
Releases: main, 12.4
Change-Id: I42850eddf788fabf04e8f51b88e4f053d947a63a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83418
Tested-by: core-ci <typo3@b13.com>
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Lina Wolf <112@linawolf.de>