[BUGFIX] Stop event propagation when clicking icon in page tree
With #103197, a loading indicator was added to the context menus.
However, due to the nature how the context menu component registers
itself, we run into conflicting event handling. We have to silent any
subsequent handlers on the same event, therefore the event propagation
is immediately stopped.

This aligns the event propagation behavior with the modernized tree in
TYPO3 v13.

Resolves: #103379
Related: #103197
Releases: 12.4
Change-Id: I87ae7ac815d3271788a164258ab2208fe5e57e6e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83441
Tested-by: Sven Juergens <typo3@blue-side.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>