[TASK] Add description to tableRestrictions palette in BE groups
This commit introduces a description for the newly added
tableRestrictions palette in the Backend (BE) user groups form,
which groups the tables_select and tables_modify fields together.

The description clarifies that tables designated for modification
are automatically considered as tables for listing as well.

This enhancement aims to improve user experience by eliminating
the redundant step of setting tables for listing when they have already
been marked for modification.

Resolves: #103426
Related: #103413
Releases: main
Change-Id: I0410f5d406e105f83e5cbaf13eec6322d4f74233
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83522
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>