[TASK] linkvalidator: Calculate option toggle state on render time
This commit changes how the "Toggle all" and option and action button
states are determined. Previously, the state was determined via
JavaScript, leading to an unexpected rendering change when the module
has loaded.

To avoid this, the state is now calculated in the controller as the
necessary metrics are already known.

In the same run, `optionByType.checked` is now a simple bool flag
instead of holding arbitrary HTML attributes.

Resolves: #103454
Releases: main
Change-Id: I2da39063dbb9161873274dadddcc3b0384e36321
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83560
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>