[TASK] Update core-testing-php image versions
New core-testing-* images have been build and published in
the TYPO3 testing-infrastructure repository and pulled to
the core ci gitlab runner instances. [1]

This change modifies `Build/Scripts/runTests.sh` to set
new image version to ensure the usage for all changes
based on this change.

For information what changed in the image see the patches
in the testing-infrastrucure repository linked below.

* core-testing-php82 => 1.12
* core-testing-php83 => 1.13

[1] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/ebdaefdb54a4fe970abf0cb1d32567bbd6339de5

Resolves: #103469
Releases: main, 12.4, 11.5
Change-Id: I35ffd10e7979d4fd59aab5a7fa5d52330e71e0e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83569
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>