[FEATURE] Introduce site sets with setting definitions
Site sets ship parts of site configuration as composable pieces. They
are intended to deliver settings, TypoScript, TSConfig and reference
enabled content blocks for the scope of a site.

Extensions can provide multiple sets in order to ship presets for
different sites or subsets (think of frameworks) where selected features
are exposed as a subset (example: `typo3/seo-xml-sitemap`).

A set is defined in an extensions subfolder in
:file:`Configuration/Sets/`, for example
:file:`EXT:my_extension/Configuration/Sets/MySet/config.yaml`.

The folder name in :file:`Configuration/Sets/` is arbitrary, significant
is the `name` defined in :file:`config.yaml`. The `name` uses a
`vendor/name` scheme by convention, and *should* use the same vendor as
the containing extension. It may differ if needed for compatibility
reasons (e.g. when sets are moved to other extensions).

TypoScript and PageTS providers will be added in subsequent changes
(#103439, #103522).

Technical notes:

 * Sets whose dependencies can not be fullfilled (are unavailable)
   will log an error in the TYPO3 error log and will be ignored.
   This is to ensure a broken set of one site does not break another
   site, if their dependencies can be resolved.

 * All settings interfaces are marked @internal for now, since they
   may change until v13 LTS. It is planned to remove the @internal
   annotation prior to the LTS release.

Releases: main
Resolves: #103437
Related: #103439
Related: #103522
Change-Id: I01cf60a837a69ea2216c9d8cd5ec9fbacacb5ece
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82191
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>