[TASK] Avoid calling deprecated Symfony Console helper commands
`Console\Helper::strlen()` was deprecated in Symfony 5.3. As we
require Symfony >= 5.4, we can safely use the replacement
function instead that 12.4 and `main` also use.

Note: This already got fixed in main and 12.4 with the upgrade to
Symfony 6.1. So this fix here is for 11.5 only.

Another note: This change aligns the code with 12.4 and `main`. Changing this to the more fitting `::width()` should happen in a separate later change across versions, not only for 11.5.

Resolves: #103503
Related: #98158
Releases: 11.5
Change-Id: Iabd9bf2423501c69e9cd805a6d923f779c8ca04b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83625
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>